requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
et-xmlfile==1.1.0
ffmpeg-python==0.2.0
future==1.0.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.3
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
protobuf==5.26.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tzdata==2023.4
urllib3==2.2.1
